Migrating `ExportExtractor` to TS to fix linter hints.
